Gh12202 maintaine address name (#12203)
* Maintaine address name

* Maintaine address name

* fixes/improvements:

move code exclusively used by C_BPartner_Location near C_BPartner_Location..no point in having it in MBPartnerLocation
